requirements.txt
Django==5.0
Pillow==12.0.0
django-crispy-forms==2.5
crispy-bootstrap5==2025.6
django-environ==0.11.2
whitenoise==6.6.0
django-ckeditor==6.7.0
django-storages==1.14.2
gunicorn==21.2.0
psycopg2-binary==2.9.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-decouple==3.8